+++ src/frontend/apps/impress/src/core/config/api/useConfig.tsx
@@ -4,13 +4,14 @@ import { Resource } from 'i18next';
 import { APIError, errorCauses, fetchAPI } from '@/api';
 import { Theme } from '@/cunningham/';
 import { FooterType } from '@/features/footer';
-import { HeaderType } from '@/features/header';
+import { HeaderType, WaffleType } from '@/features/header';
 import { PostHogConf } from '@/services';
 
 interface ThemeCustomization {
   footer?: FooterType;
   translations?: Resource;
   header?: HeaderType;
+  waffle?: WaffleType;
 }
 
 export interface ConfigResponse {