+++ src/frontend/apps/impress/src/features/header/components/index.ts
@@ -1,4 +1,4 @@
 export * from './ButtonTogglePanel';
 export * from './Header';
-export * from './LaGaufre';
+export * from './Waffle';
 export * from './Title';